Build on a toolchain that has the types the source uses
The first automatic run failed everywhere, which is the point of having
it: CI was pinned to Swift 6.1.2, and Garuda has used `Span` for a borrowed
body since long before now. Nobody found out, because nobody pressed the
button.

Linux is pinned to 6.3.3, the toolchain the tests are run on. macOS takes
what its newest Xcode gives, which is 6.2.4 today, and two integer
literals that 6.3 infers as UInt8 and 6.2 does not are spelled out. The
package now declares tools version 6.2 and the README says 6.2, because
that is the truth: `Span` arrived there.

+++ Sources/Garuda/TestWebSocket.swift
@@ -132,7 +132,7 @@ public final class TestWebSocket {
         if n < 126 {
             frame.append(0x80 | UInt8(n))
         } else if n <= 0xFFFF {
-            frame += [0x80 | 126, UInt8(n >> 8), UInt8(n & 0xFF)]
+            frame += [UInt8(0x80 | 126), UInt8(n >> 8), UInt8(n & 0xFF)]
         } else {
             frame.append(0x80 | 127)
             for shift in stride(from: 56, through: 0, by: -8) { frame.append(UInt8((n >> shift) & 0xFF)) }
